feat: 게시글 기부 확정 - 게시글 만료 로직 위치 변경 - 기부 확정 테스트용 controller - 기부 확정시 레드카드 상태 사용가능으로 변경
