+++ backend/src/main/java/com/redbox/domain/donation/controller/DonationController.java
@@ -2,6 +2,7 @@ package com.redbox.domain.donation.controller;
 
 import com.redbox.domain.donation.application.DonationService;
 import com.redbox.domain.donation.application.DonationStatsService;
+import com.redbox.domain.donation.application.RequestDonationService;
 import com.redbox.domain.donation.application.UserDonationService;
 import com.redbox.domain.donation.dto.DonationRequest;
 import com.redbox.domain.donation.dto.Top5DonorResponse;
@@ -23,6 +24,7 @@ public class DonationController {
 
     private final DonationStatsService donationStatsService;
     private final UserDonationService userDonationService;
+    private final RequestDonationService requestDonationService;
 
     @PostMapping("/donate/{type}")
     public ResponseEntity<String> donate(@PathVariable String type, @RequestBody DonationRequest donationRequest) {
@@ -33,7 +35,6 @@ public class DonationController {
         }
 
         donationService.processDonation(donationRequest);
-        // TODO: message dto 에 담기
         return ResponseEntity.ok("기부 성공");
     }
 
@@ -51,7 +52,13 @@ public class DonationController {
         }
 
         donationService.cancelDonation(receiveId);
-
         return ResponseEntity.ok("기부 취소");
+    }
+
+//    기부확정 테스트용 api 잘 돌아가는지 확인해보세영
+    @GetMapping("/test/confirm/{requestId}/{receiverId}")
+    public ResponseEntity<String> testConfirm(@PathVariable long requestId, @PathVariable long receiverId) {
+        requestDonationService.donationConfirm(requestId, receiverId);
+        return ResponseEntity.ok("기부 확정");
     }
 }
