+++ backend/src/main/java/com/redbox/domain/request/application/RequestService.java
@@ -44,7 +44,6 @@ public class RequestService {
     private final RequestRepository requestRepository;
     private final LikesRepository likesRepository;
     private final UserRepository userRepository;
-//    private final RequestDonationService requestDonationService;
 
     // 현재 로그인한 사용자 정보 가져오기
     private Long getCurrentUserId() {
@@ -69,17 +68,17 @@ public class RequestService {
         return new PageResponse<>(responsePage);
     }
 
-    // 게시글 만료 처리
-    @Transactional
-    public void updateExpiredRequests() {
-        LocalDate today = LocalDate.now();
-        List<Request> expiredRequests = requestRepository.findByDonationEndDateBeforeAndProgressNot(today, RequestStatus.EXPIRED);
-        for (Request request : expiredRequests) {
-//            requestDonationService.donationConfirm(request.getRequestId(), request.getUserId());
-            request.expired();
-        }
-        requestRepository.saveAll(expiredRequests);
-    }
+//    // 게시글 만료 처리
+//    @Transactional
+//    public void updateExpiredRequests() {
+//        LocalDate today = LocalDate.now();
+//        List<Request> expiredRequests = requestRepository.findByDonationEndDateBeforeAndProgressNot(today, RequestStatus.EXPIRED);
+//        for (Request request : expiredRequests) {
+////            requestDonationService.donationConfirm(request.getRequestId(), request.getUserId());
+//            request.expired();
+//        }
+//        requestRepository.saveAll(expiredRequests);
+//    }
 
     // 게시글 등록
     @Transactional
