+++ backend/src/main/java/com/redbox/global/util/scheduler/Scheduler.java
@@ -1,5 +1,6 @@
 package com.redbox.global.util.scheduler;
 
+import com.redbox.domain.donation.application.RequestDonationService;
 import com.redbox.domain.request.application.RequestService;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
@@ -12,19 +13,19 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class Scheduler {
 
-    private final RequestService requestService;
+    private final RequestDonationService requestDonationService;
 
     @Scheduled(cron = "0 0 0 * * *") // 매일 자정 실행
     public void runBatchJob() {
         log.info("만료 게시글 업데이트(자정)");
-        requestService.updateExpiredRequests();
+        requestDonationService.updateExpiredRequests();
     }
 
     // 서버 실행시 적용(test)
     @PostConstruct
     public void runOnStartup() {
         log.info("만료 게시글 업데이트");
-        requestService.updateExpiredRequests();
+        requestDonationService.updateExpiredRequests();
     }
 
 }
